chore: move db source of truth to backend package
