+++ docs/hacking.md
@@ -2,18 +2,20 @@
 
 This document presents a few ways you might want to use Ilmomasiina outside our pre-packaged image.
 
-## Embedding
+## Embedding with state hooks and providers
 
-To embed events on your site, there are a couple options:
+To embed events on your own React frontend, you can import the `@tietokilta/ilmomasiina-client` package (available on npm).
+This package contains:
 
-### State hooks and providers
+- React hooks to fetch and manage state for the Events, SingleEvent and EditSignup routes.
+- Provider components that do the fetch and provide the state in a context.
+- Locale strings in Finnish and English for all UI needed for those routes.
+- Utility functions for formatting API objects, such as flattening and sorting signups.
 
-You can import the state hooks from `ilmomasiina-client` and write your own React frontend. There are also ready-made
-provider components that do the fetch and provide the state in a context.
+## API models
 
-### API models
-
-You can also go fully custom and only import the API models from `ilmomasiina-models` (see [API](#api)).
+You can also implement a fully custom API client by only importing the API models from `@tietokilta/ilmomasiina-models`
+(also available on npm).
 
 ## App customization
 
@@ -24,8 +26,3 @@ GitHub Action that clones this repo (as a submodule or directly), replaces `_def
 If you fork the repository and don't modify `ilmomasiina-models`,
 your modified backend or frontend should be compatible with the current unmodified versions.
 Don't forget to submit a PR if your code might be useful to others!
-
-## API
-
-To build a new API client, import `ilmomasiina-models` and use your fetch library of choice. `dist/services` has
-type definitions for the API, while `dist/models` is intended for backend use.
