+++ packages/ilmomasiina-backend/src/models/event.ts
@@ -18,13 +18,48 @@ import {
 } from "sequelize";
 
 import { PaymentMode, QuestionCreate, QuotaCreate } from "@tietokilta/ilmomasiina-models";
-import type { EventAttributes, EventLanguage } from "@tietokilta/ilmomasiina-models/dist/models";
+import type { QuestionLanguage, QuotaLanguage } from "@tietokilta/ilmomasiina-models/dist/schema";
 import config from "../config";
 import { EventValidationError } from "./errors";
 import type { Question, QuestionCreationAttributes } from "./question";
 import type { Quota, QuotaCreationAttributes } from "./quota";
 import { generateRandomId, RANDOM_ID_LENGTH } from "./randomId";
 import { jsonColumnGetter } from "./util/json";
+
+interface EventPerLanguageAttributes {
+  title: string;
+  description: string | null;
+  price: string | null;
+  location: string | null;
+  webpageUrl: string | null;
+  facebookUrl: string | null;
+  verificationEmail: string | null;
+}
+
+export interface EventLanguage extends EventPerLanguageAttributes {
+  quotas: QuotaLanguage[];
+  questions: QuestionLanguage[];
+}
+
+export interface EventAttributes extends EventPerLanguageAttributes {
+  id: string;
+  slug: string;
+  date: Date | null;
+  endDate: Date | null;
+  registrationStartDate: Date | null;
+  registrationEndDate: Date | null;
+  openQuotaSize: number;
+  category: string;
+  draft: boolean;
+  listed: boolean;
+  signupsPublic: boolean;
+  nameQuestion: boolean;
+  emailQuestion: boolean;
+  payments: PaymentMode;
+  languages: Record<string, EventLanguage>;
+  defaultLanguage: string;
+  updatedAt: Date;
+}
 
 // Drop updatedAt so we don't need to define it manually in Event.init().
 // updatedAt is in EventAttributes since it's referenced in the adminEventListEventAttrs array, which is
