+++ packages/ilmomasiina-backend/src/models/quota.ts
@@ -18,10 +18,19 @@ import {
   Sequelize,
 } from "sequelize";
 
-import type { QuotaAttributes } from "@tietokilta/ilmomasiina-models/dist/models";
 import type { Event } from "./event";
 import { generateRandomId, RANDOM_ID_LENGTH } from "./randomId";
 import type { Signup } from "./signup";
+
+export interface QuotaAttributes {
+  id: string;
+  order: number;
+  title: string;
+  size: number | null;
+  eventId: Event["id"];
+  signupCount?: number;
+  price: number;
+}
 
 export interface QuotaCreationAttributes extends Optional<QuotaAttributes, "id"> {}
