+++ packages/ilmomasiina-backend/src/models/signup.ts
@@ -24,13 +24,32 @@ import {
 } from "sequelize";
 
 import { ProductSchema, SignupStatus } from "@tietokilta/ilmomasiina-models";
-import type { SignupAttributes } from "@tietokilta/ilmomasiina-models/dist/models";
 import config from "../config";
 import type { Answer } from "./answer";
 import type { Payment } from "./payment";
 import type { Quota } from "./quota";
 import { generateRandomId, RANDOM_ID_LENGTH } from "./randomId";
 import { jsonColumnGetter } from "./util/json";
+
+export interface SignupAttributes {
+  id: string;
+  firstName: string | null;
+  lastName: string | null;
+  namePublic: boolean;
+  email: string | null;
+  language: string | null;
+  confirmedAt: Date | null;
+  status: SignupStatus | null;
+  position: number | null;
+  /** Total price of the signup in cents, calculated when it was last updated. */
+  price: number | null;
+  /** The currency in which the price is denominated. */
+  currency: string | null;
+  /** The product lines used to calculate the price. */
+  products: ProductSchema[] | null;
+  createdAt: Date;
+  quotaId: Quota["id"];
+}
 
 export interface SignupCreationAttributes extends Optional<
   SignupAttributes,
