+++ packages/ilmomasiina-backend/src/models/user.ts
@@ -1,6 +1,10 @@
 import { DataTypes, Model, Optional, Sequelize } from "sequelize";
 
-import type { UserAttributes } from "@tietokilta/ilmomasiina-models/dist/models";
+export interface UserAttributes {
+  id: number;
+  email: string;
+  password: string;
+}
 
 export interface UserCreationAttributes extends Optional<UserAttributes, "id"> {}
