--- packages/ilmomasiina-models/src/models/answer.ts
@@ -1,9 +1,0 @@
-import type QuestionAttributes from "./question";
-import type SignupAttributes from "./signup";
-
-export default interface AnswerAttributes {
-  id: string;
-  answer: string | string[];
-  questionId: QuestionAttributes["id"];
-  signupId: SignupAttributes["id"];
-}
