--- packages/ilmomasiina-models/src/models/index.ts
@@ -1,8 +1,0 @@
-export type { default as AnswerAttributes } from "./answer";
-export type { default as AuditLogAttributes } from "./auditlog";
-export type { default as EventAttributes, EventLanguage } from "./event";
-export type { default as PaymentAttributes } from "./payment";
-export type { default as QuestionAttributes } from "./question";
-export type { default as QuotaAttributes } from "./quota";
-export type { default as SignupAttributes } from "./signup";
-export type { default as UserAttributes } from "./user";
