--- packages/ilmomasiina-models/src/models/quota.ts
@@ -1,11 +1,0 @@
-import type EventAttributes from "./event";
-
-export default interface QuotaAttributes {
-  id: string;
-  order: number;
-  title: string;
-  size: number | null;
-  eventId: EventAttributes["id"];
-  signupCount?: number;
-  price: number;
-}
