--- packages/ilmomasiina-models/src/models/user.ts
@@ -1,5 +1,0 @@
-export default interface UserAttributes {
-  id: number;
-  email: string;
-  password: string;
-}
