+++ packages/ilmomasiina-models/src/schema/signup/attributes.ts
@@ -1,6 +1,6 @@
 import { Type } from "typebox";
 
-import { PaymentStatus, SignupStatus } from "../../enum";
+import { SignupPaymentStatus, SignupStatus } from "../../enum";
 import { productSchema } from "../product";
 import { questionID } from "../question/attributes";
 import { Nullable } from "../utils";
@@ -102,7 +102,7 @@ export const adminDynamicSignupAttributes = Type.Interface([publicDynamicSignupA
   currency: Nullable(Type.String({ maxLength: 8 }), {
     description: "The currency in which the price is denominated.",
   }),
-  paymentStatus: Nullable(Type.Enum(PaymentStatus), {
+  paymentStatus: Nullable(Type.Enum(SignupPaymentStatus), {
     description: "Status of the payment for the signup, if applicable.",
   }),
 });
